Add files via upload
Write an application for the Summerdale Condo Sales office; the program
determines the price of a condominium. Ask the user to choose 1 for park view, 2
for golf course view, or 3 for lake view. The output is the name of the chosen view
as well as the price of the condo. Park view condos are $150,000, condos with
golf course views are $170,000, and condos with lake views are $210,000. If the
user enters an invalid code, set the price to 0.
b. Add a prompt to the CondoSales application to ask the user to specify a (1)
garage or a (2) parking space, but only if the view selection is valid. Add $5,000
to the price of any condo with a garage. If the parking value is invalid, display an
appropriate message and assume that the price is for a condo with no garage.
Save the file as LifeLeisure